Working Version 1
Random Temperature values in the range [-20, -20] to display the list of
temperatures.
Sensor Manager for measuring ambient temperature of the device.
Restore display of the list choices, on change of display orientation
Change Celsius to Fahrenheit on toggle of choice
Reset button to re-spawn list view

To be done:
UI a little messed up when rotated
Random temperature generation doesnt take care of choice of units